+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 * Breaking change: The new `UserExceptionDialog.useLocalContext` parameter now allows
   the `UserExceptionDialog` to be put in the `builder` parameter of the `MaterialApp` widget. Even
-  if you use this dialog, it is unlikely this will be a breaking change for you. But if it is (your
-  error dialog now has problems), simply make `useLocalContext: true` to return to the old behavior.
+  if you use this dialog, it is unlikely this will be a breaking change for you. But if it is, and
+  your error dialog now has problems, simply make `useLocalContext: true` to return to the old
+  behavior.
+
+* Breaking change: `StoreConnector` parameters like `onInitialBuild`, `onDidChange`
+  and `.onWillChange` now also get the context and the store. For example, where you
+  previously had `onInitialBuild(vm) {...}` now you have `onInitialBuild(context, store, vm) {...}`.
 
 # [9.0.9] - 2021/05/10
 
@@ -87,7 +92,7 @@
 
 # [5.0.0] - 2020/10/19
 
-* Breaking change: OnWillChangeCallback now provides previousViewModel.
+* Breaking change: OnWillChangeCallback now provides previousVm.
 
 # [4.0.4] - 2020/10/19